refactor(search applications): rename schema file
